Refactor main.ts to move search request and normal mode into search.ts
